Add files via upload
When I ran the marro in the test workbook it came out as we wanted, although it did require clicking the macro button twice it came out nicely. When I ran it in the main workbook however it did not run as nicely, which is why I attached the test workbook as well to show my work. Thanks